requirements.txt
django
djangorestframework
pytest-django
factory_boy
Faker
responses
celery
